Move some function to utils
