+++ dask_kubernetes/utils.py
@@ -1,4 +1,5 @@
 """Utility functions."""
+import asyncio
 import os
 import random
 import shutil
@@ -9,6 +10,7 @@ import time
 from weakref import finalize
 
 from dask.distributed import Client
+import kubernetes_asyncio as kubernetes
 
 
 def format_labels(labels):
@@ -134,3 +136,29 @@ def check_dependency(dependency):
             f"Missing dependency {dependency}. "
             f"Please install {dependency} following the instructions for your OS. "
         )
+
+
+async def get_scheduler_address(service_name, namespace):
+    async with kubernetes.client.api_client.ApiClient() as api_client:
+        api = kubernetes.client.CoreV1Api(api_client)
+        service = await api.read_namespaced_service(service_name, namespace)
+        port_forward_cluster_ip = None
+        address = await get_external_address_for_scheduler_service(
+            api, service, port_forward_cluster_ip=port_forward_cluster_ip
+        )
+        return address
+
+
+async def wait_for_scheduler(cluster_name, namespace):
+    async with kubernetes.client.api_client.ApiClient() as api_client:
+        api = kubernetes.client.CoreV1Api(api_client)
+        watch = kubernetes.watch.Watch()
+        async for event in watch.stream(
+            func=api.list_namespaced_pod,
+            namespace=namespace,
+            label_selector=f"dask.org/cluster-name={cluster_name},dask.org/component=scheduler",
+            timeout_seconds=60,
+        ):
+            if event["object"].status.phase == "Running":
+                watch.stop()
+            await asyncio.sleep(0.1)
